+++ cluster_utils/communication_server.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import asyncio
-import enum
 import logging
 import pickle
 import signal
@@ -9,18 +8,10 @@ import socket
 import threading
 import time
 
-from . import constants
+from cluster_utils.base import constants
+from cluster_utils.base.communication import MessageTypes
+
 from .job import JobStatus
-
-
-class MessageTypes(enum.IntEnum):
-    JOB_STARTED = 0
-    ERROR_ENCOUNTERED = 1
-    JOB_SENT_RESULTS = 2
-    JOB_CONCLUDED = 3
-    EXIT_FOR_RESUME = 4
-    JOB_PROGRESS_PERCENTAGE = 5
-    METRIC_EARLY_REPORT = 6
 
 
 class DatagramProtocol:
