+++ cluster_utils/utils.py
@@ -21,7 +21,7 @@ from typing import Any
 
 import colorama
 
-from . import constants
+from cluster_utils.base import constants
 
 
 class OptionalDependencyNotFoundError(ModuleNotFoundError):
@@ -139,19 +139,6 @@ def rm_dir_full(dir_name):
 
     if os.path.exists(dir_name):
         logger.warning(f"Removing of dir {dir_name} failed")
-
-
-def flatten_nested_string_dict(nested_dict, prepend=""):
-    for key, value in nested_dict.items():
-        if not isinstance(key, str):
-            raise TypeError("Only strings as keys expected")
-        if isinstance(value, dict):
-            for sub in flatten_nested_string_dict(
-                value, prepend=prepend + str(key) + constants.OBJECT_SEPARATOR
-            ):
-                yield sub
-        else:
-            yield prepend + str(key), value
 
 
 def get_sample_generator(
